Move methods and fix tests.
